requirements.txt
asgiref==3.6.0
Django==3.2
djangorestframework==3.12.4
pytz==2023.3
sqlparse==0.4.3
djoser==2.1.0
djangorestframework-simplejwt
Pillow==9.0.0